+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 
         -   `<XXX elevation="none">` — Removes elevation from the Component.
 
+    -   Display
+
+        -   `Table`
+
+            -   `<Table.Container sizing="tiny/small/medium/large/huge">` — Added support for modifying the text sizing / spacings of all children.
+
     -   Overlays
 
         -   `Overlay`